Style update - uint vs uint256 (#1076)

+++ smart-contracts/contracts/interfaces/IERC721.sol
@@ -12,13 +12,13 @@ interface IERC721 {
   ///  (`to` == 0). Exception: during contract creation, any number of NFTs
   ///  may be created and assigned without emitting Transfer. At the time of
   ///  any transfer, the approved address for that NFT (if any) is reset to none.
-  event Transfer(address indexed _from, address indexed _to, uint256 indexed _tokenId);
+  event Transfer(address indexed _from, address indexed _to, uint indexed _tokenId);
 
   /// @dev This emits when the approved address for an NFT is changed or
   ///  reaffirmed. The zero address indicates there is no approved address.
   ///  When a Transfer event emits, this also indicates that the approved
   ///  address for that NFT (if any) is reset to none.
-  event Approval(address indexed _owner, address indexed _approved, uint256 indexed _tokenId);
+  event Approval(address indexed _owner, address indexed _approved, uint indexed _tokenId);
 
   /// @dev This emits when an operator is enabled or disabled for an owner.
   ///  The operator can manage all NFTs of the owner.
@@ -34,7 +34,7 @@ interface IERC721 {
   /// @param _from The current owner of the NFT
   /// @param _to The new owner
   /// @param _tokenId The NFT to transfer
-  function transferFrom(address _from, address _to, uint256 _tokenId) external payable;
+  function transferFrom(address _from, address _to, uint _tokenId) external payable;
 
   /// @notice Set or reaffirm the approved address for an NFT
   /// @dev The zero address indicates there is no approved address.
@@ -42,21 +42,21 @@ interface IERC721 {
   ///  operator of the current owner.
   /// @param _approved The new approved NFT controller
   /// @param _tokenId The NFT to approve
-  function approve(address _approved, uint256 _tokenId) external payable;
+  function approve(address _approved, uint _tokenId) external payable;
 
   /// @notice Count all NFTs assigned to an owner
   /// @dev NFTs assigned to the zero address are considered invalid, and this
   ///  function throws for queries about the zero address.
   /// @param _owner An address for whom to query the balance
   /// @return The number of NFTs owned by `_owner`, possibly zero
-  function balanceOf(address _owner) external view returns (uint256);
+  function balanceOf(address _owner) external view returns (uint);
 
   /// @notice Find the owner of an NFT
   /// @dev NFTs assigned to zero address are considered invalid, and queries
   ///  about them do throw.
   /// @param _tokenId The identifier for an NFT
   /// @return The address of the owner of the NFT
-  function ownerOf(uint256 _tokenId) external view returns (address);
+  function ownerOf(uint _tokenId) external view returns (address);
 
   /// @notice Transfers the ownership of an NFT from one address to another address
   /// @dev Throws unless `msg.sender` is the current owner, an authorized
@@ -65,12 +65,12 @@ interface IERC721 {
   ///  `_tokenId` is not a valid NFT. When transfer is complete, this function
   ///  checks if `_to` is a smart contract (code size > 0). If so, it calls
   ///  `onERC721Received` on `_to` and throws if the return value is not
-  ///  `bytes4(keccak256("onERC721Received(address,address,uint256,bytes)"))`.
+  ///  `bytes4(keccak256("onERC721Received(address,address,uint,bytes)"))`.
   /// @param _from The current owner of the NFT
   /// @param _to The new owner
   /// @param _tokenId The NFT to transfer
   /// @param data Additional data with no specified format, sent in call to `_to`
-  // function safeTransferFrom(address _from, address _to, uint256 _tokenId, bytes data) external payable;
+  // function safeTransferFrom(address _from, address _to, uint _tokenId, bytes data) external payable;
 
   /// @notice Transfers the ownership of an NFT from one address to another address
   /// @dev This works identically to the other function with an extra data parameter,
@@ -78,7 +78,7 @@ interface IERC721 {
   /// @param _from The current owner of the NFT
   /// @param _to The new owner
   /// @param _tokenId The NFT to transfer
-  // function safeTransferFrom(address _from, address _to, uint256 _tokenId) external payable;
+  // function safeTransferFrom(address _from, address _to, uint _tokenId) external payable;
 
   /// @notice Enable or disable approval for a third party ("operator") to manage
   ///  all of `msg.sender`'s assets.
@@ -92,7 +92,7 @@ interface IERC721 {
   /// @dev Throws if `_tokenId` is not a valid NFT
   /// @param _tokenId The NFT to find the approved address for
   /// @return The approved address for this NFT, or the zero address if there is none
-  function getApproved(uint256 _tokenId) external view returns (address);
+  function getApproved(uint _tokenId) external view returns (address);
 
   /// @notice Query if an address is an authorized operator for another address
   /// @param _owner The address that owns the NFTs
@@ -111,12 +111,12 @@ interface IERC721 {
 //     /// @param _from The address which previously owned the token
 //     /// @param _tokenId The NFT identifier which is being transferred
 //     /// @param _data Additional data with no specified format
-//     /// @return `bytes4(keccak256("onERC721Received(address,address,uint256,bytes)"))`
+//     /// @return `bytes4(keccak256("onERC721Received(address,address,uint,bytes)"))`
 //     /// unless throwing
     // function onERC721Received(
     //   address _operator,
     //   address _from,
-    //   uint256 _tokenId,
+    //   uint _tokenId,
     //   bytes _data
     // )
     // external
